+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Fixed
+- `genomepy clean` wont complain when there is nothing to clean
+
 ## [0.9.0] - 2020-09-01
 
 ### Added